Fixed 64 bit issues

+++ WebServices/SDWebService+SDProcessingBlocks.m
@@ -108,7 +108,7 @@
     NSAssert([classType isSubclassOfClass:[SDModelObject class]], @"defaultJSONProcessingBlockForClass: works on concrete subclasses of SDModelObject");
 
     SDWebServiceDataCompletionBlock completionBlock = ^id (NSURLResponse *response, NSInteger responseCode, NSData *responseData, NSError *error) {
-        SDLog(@"%@: %d:\n%@", response, responseCode, [[NSString alloc] initWithData:responseData encoding:NSUTF8StringEncoding]);
+        SDLog(@"%@: %d:\n%@", response, (int)responseCode, [[NSString alloc] initWithData:responseData encoding:NSUTF8StringEncoding]);
 
         id responseObject = [responseData JSONObject];
